[Version Number Comparison] Change: Extract to new file
